Actually execute asserts during replay
During replay, asserts are executed and evaluated now using the
Connection message 'executeAssert'. This is also executed during
test-step generation.

The following steps are needed to implement the assert execution during
replay:
- The functionality to execute asserts is moved to the page injection
and is carried out on the Connection message 'executeAssert'. The result
of the execution is analogous to the result of of the messages
'executeAction' and 'checkAction'.
- The controller 'TestStep' now sends Connection messages for checking
actions and executing asserts and collects the incoming messsages.
- The methods '(_)getAssertDefinition' and '(_)getValueSpec' are moved
from the controller 'TestStep' to 'Utils'. (More methods are likely to
come!)
- In the same vein, criteria types, attribute types, and element mix are
moved from 'GlobalSettings' to 'Utils' as they are rather helper
structures than settings. The 'Utils' are now – analogously to the
'GlobalSettings' – also a sap.ui.base.Object to ensure proper storage.
